fix: separate powerspy and powerjoular execution

+++ scripts/raspi/stop_all.sh
@@ -2,6 +2,8 @@
 
 set -euo pipefail
 
-kill $(cat ../../powerexp/raspi/pids.txt)
+if [ -f ~/ERC-4337-Bundlers-Test-Setup/powerexp/raspi/pids.txt ]; then
+	kill $(cat ../../powerexp/raspi/pids.txt)
+fi
 
 sudo docker stop $(sudo docker ps -a -q)
